requirements.txt
opencv-python
face_recognition
pandas
numpy
pillow